requirements-docs.txt
# Requirements for building the documentation
sphinx
sphinx_gallery
sphinx_bootstrap_theme
sphinx-copybutton
numpydoc

# Requirements for running the examples
mne
pillow
